requirements.txt
python-telegram-bot[ext]==20.7
sqlalchemy[asyncio]==2.0.23
aiomysql==0.2.0
cryptography==41.0.7
python-dotenv==1.0.0
Pillow==10.2.0
